segtrain/requirement.txt
tensorflow>=2.13
keras>=2.13
numpy
pillow
opencv-python
scikit-image
scipy
pydensecrf @ git+https://github.com/lucasb-eyer/pydensecrf.git
